refactor:#265 client set-goal 리팩토링
