+++ src/app/(client)/set-goal/_components/step-ai-loading.tsx
