+++ src/app/(client)/set-goal/_components/step-weight.tsx
