+++ src/app/(client)/set-goal/_hooks/use-scroll-to-top.ts
