+++ src/app/(client)/set-goal/_types/funnel.type.ts
